+++ libc/stdio/printf.c
@@ -160,7 +160,7 @@ int printf(const char* restrict format, ...) {
     
         }else if(*format == 'x'){
             format++;
-            uint64_t hex = va_arg(parameters, uint64_t);
+            long long hex = va_arg(parameters, uint64_t);
 
             printf("%s", itoa(hex, &buf[0], 16));
